Separating usb callbacks from main routine
