+++ firmware/main.c
@@ -20,11 +20,10 @@
  */
 
 #include "usb.h"
+#include "usb/usb_callbacks.c"
+
 #include <xc.h>
 #include <string.h>
-#include "usb_config.h"
-#include "usb_ch9.h"
-#include "usb_cdc.h"
 
 #include "io.h"
 #include "parse.h"
@@ -305,141 +304,6 @@ empty:
 
     return 0;
 }
-
-/* Callbacks. These function names are set in usb_config.h. */
-void app_set_configuration_callback(uint8_t configuration)
-{
-
-}
-
-uint16_t app_get_device_status_callback()
-{
-    return 0x0000;
-}
-
-void app_endpoint_halt_callback(uint8_t endpoint, bool halted)
-{
-
-}
-
-int8_t app_set_interface_callback(uint8_t interface, uint8_t alt_setting)
-{
-    return 0;
-}
-
-int8_t app_get_interface_callback(uint8_t interface)
-{
-    return 0;
-}
-
-void app_out_transaction_callback(uint8_t endpoint)
-{
-
-}
-
-void app_in_transaction_complete_callback(uint8_t endpoint)
-{
-
-}
-
-int8_t app_unknown_setup_request_callback(const struct setup_packet *setup)
-{
-    /* To use the CDC device class, have a handler for unknown setup
-     * requests and call process_cdc_setup_request() (as shown here),
-     * which will check if the setup request is CDC-related, and will
-     * call the CDC application callbacks defined in usb_cdc.h. For
-     * composite devices containing other device classes, make sure
-     * MULTI_CLASS_DEVICE is defined in usb_config.h and call all
-     * appropriate device class setup request functions here.
-     */
-    return process_cdc_setup_request(setup);
-}
-
-int16_t app_unknown_get_descriptor_callback(const struct setup_packet *pkt, const void **descriptor)
-{
-    return -1;
-}
-
-void app_start_of_frame_callback(void)
-{
-
-}
-
-void app_usb_reset_callback(void)
-{
-
-}
-
-/* CDC Callbacks. See usb_cdc.h for documentation. */
-
-int8_t app_send_encapsulated_command(uint8_t interface, uint16_t length)
-{
-    return -1;
-}
-
-int16_t app_get_encapsulated_response(uint8_t interface,
-                                      uint16_t length, const void **report,
-                                      usb_ep0_data_stage_callback *callback,
-                                      void **context)
-{
-    return -1;
-}
-
-int8_t app_set_comm_feature_callback(uint8_t interface,
-                                     bool idle_setting,
-                                     bool data_multiplexed_state)
-{
-    return -1;
-}
-
-int8_t app_clear_comm_feature_callback(uint8_t interface,
-                                       bool idle_setting,
-                                       bool data_multiplexed_state)
-{
-    return -1;
-}
-
-int8_t app_get_comm_feature_callback(uint8_t interface,
-                                     bool *idle_setting,
-                                     bool *data_multiplexed_state)
-{
-    return -1;
-}
-
-static struct cdc_line_coding line_coding =
-{
-    115200,
-    CDC_CHAR_FORMAT_1_STOP_BIT,
-    CDC_PARITY_NONE,
-    8,
-};
-
-int8_t app_set_line_coding_callback(uint8_t interface,
-                                    const struct cdc_line_coding *coding)
-{
-    line_coding = *coding;
-    return 0;
-}
-
-int8_t app_get_line_coding_callback(uint8_t interface,
-                                    struct cdc_line_coding *coding)
-{
-    /* This is where baud rate, data, stop, and parity bits are set. */
-    *coding = line_coding;
-    return 0;
-}
-
-int8_t app_set_control_line_state_callback(uint8_t interface,
-                                           bool dtr, bool dts)
-{
-    return 0;
-}
-
-int8_t app_send_break_callback(uint8_t interface, uint16_t duration)
-{
-    return 0;
-}
-
 
 void interrupt high_priority isr()
 {
